pyproject.toml
[project]
name = "superset-warmer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest-playwright>=0.7.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]
